fixed test Makefiles

+++ test/am_epg_test/am_epg_test.c
@@ -279,13 +279,13 @@ static int start_epg_test()
 	disable_defproc_reg_cb(hmon);
 
 	/*注册通知事件*/
-	AM_EVT_Subscribe(hmon, AM_EPG_EVT_NEW_PAT, epg_evt_call_back, NULL);
-	AM_EVT_Subscribe(hmon, AM_EPG_EVT_NEW_PMT, epg_evt_call_back, NULL);
-	AM_EVT_Subscribe(hmon, AM_EPG_EVT_NEW_CAT, epg_evt_call_back, NULL);
-	AM_EVT_Subscribe(hmon, AM_EPG_EVT_NEW_SDT, epg_evt_call_back, NULL);
-	AM_EVT_Subscribe(hmon, AM_EPG_EVT_NEW_NIT, epg_evt_call_back, NULL);
-	AM_EVT_Subscribe(hmon, AM_EPG_EVT_NEW_TDT, epg_evt_call_back, NULL);
-	AM_EVT_Subscribe(hmon, AM_EPG_EVT_NEW_EIT, epg_evt_call_back, NULL);
+	AM_EVT_Subscribe((size_t)hmon, AM_EPG_EVT_NEW_PAT, epg_evt_call_back, NULL);
+	AM_EVT_Subscribe((size_t)hmon, AM_EPG_EVT_NEW_PMT, epg_evt_call_back, NULL);
+	AM_EVT_Subscribe((size_t)hmon, AM_EPG_EVT_NEW_CAT, epg_evt_call_back, NULL);
+	AM_EVT_Subscribe((size_t)hmon, AM_EPG_EVT_NEW_SDT, epg_evt_call_back, NULL);
+	AM_EVT_Subscribe((size_t)hmon, AM_EPG_EVT_NEW_NIT, epg_evt_call_back, NULL);
+	AM_EVT_Subscribe((size_t)hmon, AM_EPG_EVT_NEW_TDT, epg_evt_call_back, NULL);
+	AM_EVT_Subscribe((size_t)hmon, AM_EPG_EVT_NEW_EIT, epg_evt_call_back, NULL);
 
 	printf("------------------------------------------\n");
 	printf("commands:\n");
@@ -455,13 +455,13 @@ static int start_epg_test()
 	AM_TRY(AM_EPG_Destroy(hmon));
 	AM_DEBUG(1, "OK");
 
-	AM_EVT_Unsubscribe(hmon, AM_EPG_EVT_NEW_PAT, epg_evt_call_back, NULL);
-	AM_EVT_Unsubscribe(hmon, AM_EPG_EVT_NEW_PMT, epg_evt_call_back, NULL);
-	AM_EVT_Unsubscribe(hmon, AM_EPG_EVT_NEW_CAT, epg_evt_call_back, NULL);
-	AM_EVT_Unsubscribe(hmon, AM_EPG_EVT_NEW_SDT, epg_evt_call_back, NULL);
-	AM_EVT_Unsubscribe(hmon, AM_EPG_EVT_NEW_NIT, epg_evt_call_back, NULL);
-	AM_EVT_Unsubscribe(hmon, AM_EPG_EVT_NEW_TDT, epg_evt_call_back, NULL);
-	AM_EVT_Unsubscribe(hmon, AM_EPG_EVT_NEW_EIT, epg_evt_call_back, NULL);
+	AM_EVT_Unsubscribe((size_t)hmon, AM_EPG_EVT_NEW_PAT, epg_evt_call_back, NULL);
+	AM_EVT_Unsubscribe((size_t)hmon, AM_EPG_EVT_NEW_PMT, epg_evt_call_back, NULL);
+	AM_EVT_Unsubscribe((size_t)hmon, AM_EPG_EVT_NEW_CAT, epg_evt_call_back, NULL);
+	AM_EVT_Unsubscribe((size_t)hmon, AM_EPG_EVT_NEW_SDT, epg_evt_call_back, NULL);
+	AM_EVT_Unsubscribe((size_t)hmon, AM_EPG_EVT_NEW_NIT, epg_evt_call_back, NULL);
+	AM_EVT_Unsubscribe((size_t)hmon, AM_EPG_EVT_NEW_TDT, epg_evt_call_back, NULL);
+	AM_EVT_Unsubscribe((size_t)hmon, AM_EPG_EVT_NEW_EIT, epg_evt_call_back, NULL);
 
 	return 0;
 }
